--- a/src/abstractdataframe/reshape.jl
+++ b/src/abstractdataframe/reshape.jl
@@ -405,16 +405,30 @@ function transpose1(df::AbstractDataFrame, src_namescol=1, dest_namescol=:variab
     unstack(stack(df,Not(src_namescol), variable_name=dest_namescol), dest_namescol, src_namescol, :value)
 end
 
-function transpose2(df::AbstractDataFrame, src_namescol=1, dest_namescol=:variable)
-    m = permutedims((Matrix(df[!, Not(src_namescol)])))
+function transpose2(df::AbstractDataFrame, src_namescol=1, dest_namescol=:variable; copycols=false)
+    m = permutedims(Matrix(df[!, Not(src_namescol)]))
     df2 = DataFrame([names(df[!, Not(src_namescol)])], [dest_namescol])
-    hcat(df2, DataFrame(m, df[!, src_namescol]), copycols=false)
+    hcat(df2, DataFrame(m, df[!, src_namescol]), copycols=copycols)
+end
+
+function transpose2_comprehension(df::AbstractDataFrame, src_namescol=1, dest_namescol=:variable; copycols=false)
+    m = permutedims(Matrix(df[!, Not(src_namescol)]))
+    df2 = DataFrame([names(df[!, Not(src_namescol)])], [dest_namescol])
+    hcat(df2, DataFrame([[x for x in col] for col in eachcol(m)], df[!, src_namescol]), copycols=copycols)
 end
 
 function transpose3(df::AbstractDataFrame, src_namescol=1, dest_namescol=:variable)
     df2 = DataFrame([names(df[!, Not(src_namescol)])], [dest_namescol])
     for row in eachrow(df)
         df2[!,row[src_namescol]] = Vector(row[Not(src_namescol)])
+    end
+    return df2
+end
+
+function transpose3_comprehension(df::AbstractDataFrame, src_namescol=1, dest_namescol=:variable)
+    df2 = DataFrame([names(df[!, Not(src_namescol)])], [dest_namescol])
+    for row in eachrow(df)
+        df2[!,row[src_namescol]] =  [x for x in row[Not(src_namescol)]]
     end
     return df2
 end