… NEWS.md
… docs/src/man/reshaping_and_pivoting.md