Figured I'd get a jumpstart on Hacktoberfest, and this is something I've wanted in DataFrames for a longtime. This is an attempt to address #2420 

## todo
- [x] correct function signature
- [x] docs
- [x] tests

## So far

I implemented 3 versions of `transpose`:

- `transpose1` copies @tbeason 's stack/unstack method from the issue above
- `transpose2` creates a `Matrix` from everything _except_ the indexing column, and does `permutedims` on it
- `transpose3` builds column-by-column by iterating rows

## Benchmarking

```
julia> builddf(n) = DataFrame(a = [randstring(6) for _ in 1:n], b = rand(n), c = rand(Int, n), d = repeat(["x", "y"], inner=n÷2))
builddf (generic function with 1 method)

julia> df1 = builddf(10);

julia> df2 = builddf(100);

julia> df3 = builddf(1000);

julia> @btime DataFrames.transpose1($df1);
  33.565 μs (398 allocations: 32.61 KiB)

julia> @btime DataFrames.transpose2($df1);
  12.637 μs (179 allocations: 17.81 KiB)

julia> @btime DataFrames.transpose3($df1);
  14.723 μs (172 allocations: 16.58 KiB)

julia> @btime DataFrames.transpose1($df2);
  144.708 μs (1083 allocations: 115.92 KiB)

julia> @btime DataFrames.transpose2($df2);
  50.667 μs (650 allocations: 69.75 KiB)

julia> @btime DataFrames.transpose3($df2);
  137.353 μs (1261 allocations: 119.12 KiB)

julia> @btime DataFrames.transpose1($df3);
  1.168 ms (7924 allocations: 976.30 KiB)

julia> @btime DataFrames.transpose2($df3);
  469.847 μs (5678 allocations: 643.34 KiB)

julia> @btime DataFrames.transpose3($df3);
  1.513 ms (20035 allocations: 1.27 MiB)
```

## First impressions

- `transpose1` (stack/unstack) is slower on small `DataFrame`s but scales well.
     - This also has the possible disadvantage that the columns end up sorted, rather than preserving the row order
- `transpose2` is the fastest, but might be harder to fix type issues (see below)
- `transpose3` is pretty fast for small `DataFrame`s but doesn't scale well at all

## Other thoughts

All 3 versions end up with `Any` typed columns, regardless of what's in the resulting column (unless all the columns are the same type). This seems like it might be more easily fixed with 1 and 3. The behavior is also not quite what I was expecting if all of the columns are numerical:

```
julia> df_types = DataFrame(a=[:x, :y, :z], b=rand(3), c=["foo", 2.2, 1])
3×3 DataFrame
│ Row │ a      │ b        │ c   │
│     │ Symbol │ Float64  │ Any │
├─────┼────────┼──────────┼─────┤
│ 1   │ x      │ 0.633028 │ foo │
│ 2   │ y      │ 0.388354 │ 2.2 │
│ 3   │ z      │ 0.134275 │ 1   │

julia> DataFrames.transpose1(df_types)
2×4 DataFrame
│ Row │ variable │ x        │ y        │ z        │
│     │ String   │ Any      │ Any      │ Any      │
├─────┼──────────┼──────────┼──────────┼──────────┤
│ 1   │ b        │ 0.633028 │ 0.388354 │ 0.134275 │
│ 2   │ c        │ foo      │ 2.2      │ 1        │

julia> DataFrames.transpose2(df_types)
2×4 DataFrame
│ Row │ variable │ x        │ y        │ z        │
│     │ String   │ Any      │ Any      │ Any      │
├─────┼──────────┼──────────┼──────────┼──────────┤
│ 1   │ b        │ 0.633028 │ 0.388354 │ 0.134275 │
│ 2   │ c        │ foo      │ 2.2      │ 1        │

julia> DataFrames.transpose3(df_types)
2×4 DataFrame
│ Row │ variable │ x        │ y        │ z        │
│     │ String   │ Any      │ Any      │ Any      │
├─────┼──────────┼──────────┼──────────┼──────────┤
│ 1   │ b        │ 0.633028 │ 0.388354 │ 0.134275 │
│ 2   │ c        │ foo      │ 2.2      │ 1        │

julia> df_floats = DataFrame(a=[:x, :y, :z], b=rand(3), c=rand(3))
3×3 DataFrame
│ Row │ a      │ b        │ c        │
│     │ Symbol │ Float64  │ Float64  │
├─────┼────────┼──────────┼──────────┤
│ 1   │ x      │ 0.941859 │ 0.35769  │
│ 2   │ y      │ 0.404513 │ 0.237959 │
│ 3   │ z      │ 0.649489 │ 0.255324 │

julia> DataFrames.transpose1(df_floats)
2×4 DataFrame
│ Row │ variable │ x        │ y        │ z        │
│     │ String   │ Float64? │ Float64? │ Float64? │
├─────┼──────────┼──────────┼──────────┼──────────┤
│ 1   │ b        │ 0.941859 │ 0.404513 │ 0.649489 │
│ 2   │ c        │ 0.35769  │ 0.237959 │ 0.255324 │

julia> DataFrames.transpose2(df_floats)
2×4 DataFrame
│ Row │ variable │ x        │ y        │ z        │
│     │ String   │ Float64  │ Float64  │ Float64  │
├─────┼──────────┼──────────┼──────────┼──────────┤
│ 1   │ b        │ 0.941859 │ 0.404513 │ 0.649489 │
│ 2   │ c        │ 0.35769  │ 0.237959 │ 0.255324 │

julia> DataFrames.transpose3(df_floats)
2×4 DataFrame
│ Row │ variable │ x        │ y        │ z        │
│     │ String   │ Float64  │ Float64  │ Float64  │
├─────┼──────────┼──────────┼──────────┼──────────┤
│ 1   │ b        │ 0.941859 │ 0.404513 │ 0.649489 │
│ 2   │ c        │ 0.35769  │ 0.237959 │ 0.255324 │

julia> df_number = DataFrame(a=[:x, :y, :z], b=rand(3), c=rand(Int, 3))
3×3 DataFrame
│ Row │ a      │ b         │ c                   │
│     │ Symbol │ Float64   │ Int64               │
├─────┼────────┼───────────┼─────────────────────┤
│ 1   │ x      │ 0.438549  │ 480681029893134526  │
│ 2   │ y      │ 0.489623  │ 172366260093707646  │
│ 3   │ z      │ 0.0132915 │ 3468722734556270181 │

julia> DataFrames.transpose1(df_number)
2×4 DataFrame
│ Row │ variable │ x          │ y          │ z          │
│     │ String   │ Float64?   │ Float64?   │ Float64?   │
├─────┼──────────┼────────────┼────────────┼────────────┤
│ 1   │ b        │ 0.438549   │ 0.489623   │ 0.0132915  │
│ 2   │ c        │ 4.80681e17 │ 1.72366e17 │ 3.46872e18 │

julia> DataFrames.transpose2(df_number)
2×4 DataFrame
│ Row │ variable │ x          │ y          │ z          │
│     │ String   │ Float64    │ Float64    │ Float64    │
├─────┼──────────┼────────────┼────────────┼────────────┤
│ 1   │ b        │ 0.438549   │ 0.489623   │ 0.0132915  │
│ 2   │ c        │ 4.80681e17 │ 1.72366e17 │ 3.46872e18 │

julia> DataFrames.transpose3(df_number)
2×4 DataFrame
│ Row │ variable │ x          │ y          │ z          │
│     │ String   │ Float64    │ Float64    │ Float64    │
├─────┼──────────┼────────────┼────────────┼────────────┤
│ 1   │ b        │ 0.438549   │ 0.489623   │ 0.0132915  │
│ 2   │ c        │ 4.80681e17 │ 1.72366e17 │ 3.46872e18 │
```

Still lots to do here, but I thought it would be worth it to get the ball rolling and see if people have any thoughts. I will probably get back to this in earnest next week.