@bkamins 2020-09-21
Thank you for working on it. To get a safe eltype of the resulting columns using `typejoin` should be a safe choice I think.

The issues you raise are exactly the things that I expected/feared (i.e. that it would be hard to find the optimal solution)