@kescobo 2020-10-02
Aright, doing a bit more on this. 

1. Changed function signature and included destination col argument
2. Added some other signatures to try out. I put some benchmark code in a [gist here](https://gist.github.com/kescobo/b248b529e453e207b8bea0d43f6fd25d)

Results are for `Any` type (a), numerical types including Bool (n), and for dataframes where rows have mixed type (some all numerical, some with strings) for different numbers of rows.

This is basically the same as before.

```
┌ Warning: a10
└ @ Main ~/repos/DataFramesDev/benchmarks.jl:17
  35.006 μs (401 allocations: 32.67 KiB)
  11.873 μs (160 allocations: 16.30 KiB)
  14.772 μs (166 allocations: 16.39 KiB)
┌ Warning: a100
└ @ Main ~/repos/DataFramesDev/benchmarks.jl:21
  136.785 μs (1086 allocations: 115.98 KiB)
  46.247 μs (541 allocations: 58.39 KiB)
  143.145 μs (1255 allocations: 118.94 KiB)
┌ Warning: a1k
└ @ Main ~/repos/DataFramesDev/benchmarks.jl:25
  1.157 ms (7927 allocations: 976.36 KiB)
  428.862 μs (4669 allocations: 533.55 KiB)
  1.529 ms (20029 allocations: 1.27 MiB)
┌ Warning: n10
└ @ Main ~/repos/DataFramesDev/benchmarks.jl:29
  34.351 μs (361 allocations: 32.05 KiB)
  12.244 μs (140 allocations: 15.98 KiB)
  15.950 μs (166 allocations: 16.39 KiB)
┌ Warning: n100
└ @ Main ~/repos/DataFramesDev/benchmarks.jl:33
  132.626 μs (686 allocations: 109.73 KiB)
  45.631 μs (341 allocations: 55.27 KiB)
  151.399 μs (1255 allocations: 118.94 KiB)
┌ Warning: n1k
└ @ Main ~/repos/DataFramesDev/benchmarks.jl:37
  1.076 ms (3927 allocations: 913.86 KiB)
  429.272 μs (2669 allocations: 502.30 KiB)
  1.633 ms (20029 allocations: 1.27 MiB)
┌ Warning: mr10
└ @ Main ~/repos/DataFramesDev/benchmarks.jl:41
  33.293 μs (407 allocations: 32.77 KiB)
  12.441 μs (160 allocations: 16.30 KiB)
  15.000 μs (166 allocations: 16.39 KiB)
┌ Warning: mr100
└ @ Main ~/repos/DataFramesDev/benchmarks.jl:45
  135.343 μs (1121 allocations: 116.53 KiB)
  46.999 μs (541 allocations: 58.39 KiB)
  143.275 μs (1255 allocations: 118.94 KiB)
┌ Warning: mr1k
└ @ Main ~/repos/DataFramesDev/benchmarks.jl:49
  1.191 ms (8261 allocations: 981.58 KiB)
  438.124 μs (4669 allocations: 533.55 KiB)
  1.546 ms (20029 allocations: 1.27 MiB)
```

A couple of new observations:

1. with @nalimilan 's suggestion for `copycols` for the permuted matrix, there's a moderate speed-up. Though I had to remove the inner  `copycols` argument (for the constructor with the matrix), as it was causing a method error

```
julia> @btime DataFrames.transpose2($mr1k, copycols=false);
  425.934 μs (4669 allocations: 533.55 KiB)

julia> @btime DataFrames.transpose2($mr1k, copycols=true);
  459.766 μs (5672 allocations: 643.16 KiB)
```
With both copycols:
```
ERROR: MethodError: no method matching DataFrame(::Matrix{Any}, ::Vector{String}; copycols=true)
Closest candidates are:
  DataFrame(::AbstractMatrix{T} where T, ::AbstractVector{var"#s25"} where var"#s25"<:AbstractString; makeunique) at /home/kevin/repos/DataFramesDev/dev/DataFrames.jl/src/dataframe/dataframe.jl:272 got unsupported keyword argument "copycols"
```

2. I made a comprehension version of both `transpose2` and `transpose3` to help with types for mixed row columns. Not surprisingly, this slows down t2 (though it's still faster than t3), but I was surprised that it actually speeds up t3. That is, doing `Vector(row)` is slower than `[x for x in row]`

```
julia> @btime DataFrames.transpose2_comprehension($mr100);
  113.745 μs (1659 allocations: 116.34 KiB)

julia> @btime DataFrames.transpose2($mr100);
  48.217 μs (541 allocations: 58.39 KiB)

julia> @btime DataFrames.transpose3($mr100);
  144.665 μs (1255 allocations: 118.94 KiB)

julia> @btime DataFrames.transpose3_comprehension($mr100);
  125.924 μs (2356 allocations: 168.48 KiB)
```

Both give columns of `Real` when the row has integers and floats, and `Any` when there are `String`s in the row.

I'm not totally clear on where to add the `promote` or `typejoin` calls for `transpose2`. I tried a couple of places, and it doesn't seem to alter the output type of the columns. The promotion already seems to happen with the Matrix conversion, and can't be fixed if there are Strings in the matrix:

```
julia> n10
10×4 DataFrame
│ Row │ a      │ b         │ c                    │ d    │
│     │ String │ Float64   │ Int64                │ Bool │
├─────┼────────┼───────────┼──────────────────────┼──────┤
│ 1   │ a17Wik │ 0.356758  │ -8986863187679119824 │ 0    │
│ 2   │ os6kQD │ 0.354969  │ -1901978611141674116 │ 0    │
│ 3   │ PoiIab │ 0.769343  │ 4759929156393762244  │ 1    │
│ 4   │ 2XDnn1 │ 0.0322987 │ 2546570589132666494  │ 0    │
│ 5   │ ZIx6QI │ 0.560122  │ 6388359009763183408  │ 1    │
│ 6   │ MO1iLD │ 0.194934  │ -1409080130674345575 │ 0    │
│ 7   │ zrhAak │ 0.981582  │ -2685183505188918570 │ 0    │
│ 8   │ RZtUU1 │ 0.61182   │ -3456256066328160472 │ 1    │
│ 9   │ eoMStL │ 0.024695  │ 8468153461503185050  │ 1    │
│ 10  │ rar8je │ 0.324927  │ -983327689695525853  │ 1    │

julia> eltype(Matrix(n10[!, Not(:a)]))
Float64
```
I had some ideas with `promote`, but it throws if there are any strings encountered, and checking for mismatched types seems like it would eat any performance benefits of the permuted matrix approach.

```
julia> promote([1, 1., "str"]...)
ERROR: promotion of types Int64, Float64 and String failed to change any arguments
```

Very possible I'm missing something.