@bkamins 2020-10-02
> The promotion already seems to happen with the Matrix conversion

Yes. To avoid promotion you would have to write something like `Matrix{Any}(...)`.

So the conclusion is that `Matrix` is fastest, right? Then I would go for it and add an additional kwarg `promote_type::Bool`. If yes - then call `Matrix` (and then comprehension is not needed), if `false` then call `Matrix{Any} (and use comprehension). Just please make sure to pass `Tables.table` wrapped matrix to the constructor (cf. #2464)

In general an input from @nalimilan would be welcome on naming kwargs in this function (he is far superior than me in such things).