@kescobo 2020-10-02
> So the conclusion is that Matrix is fastest, right?

Yes, by a fair amount.

> Then I would go for it and add an additional kwarg promote_type::Bool. If yes - then call Matrix (and then comprehension is not needed), if false then call Matrix{Any} (and use comprehension).

This makes sense to me. Do you have a sense of whether the default for promotion should be `true` or `false`?