@bkamins 2020-10-02
> Do you have a sense of whether the default for promotion should be `true` or `false`?

Not sure - @nalimilan seemed to prefer `true`, but for me intuitively `false` would be safer.

Having said that I would go for `true` as I assume that most of the time when you transpose the data frame holds homogenous data in the transformed columns.