@kescobo 2020-10-02
Open questions

1. name and default for `promote_type` kwarg
2. `src_colname` and `dest_colname` should be positional, right?
3. I'm also wondering if `dest_colname` is necessary. I'd be in favor of having this be the name of the column that gets indexed on, and have users need to explicitly rename
4. Should `copycols` always be `false`? The matrix gets allocated either way, and so there's no safety issue, right?

I wrote some docs and tests, but will wait to finalize until the above are answered. Any additional feedback appreciated